Documented code. Remove unused code.
- Converted dsl to use varargs
